example(minimal-react): demo typescript path aliases
Remove obsolete npm-workspaces demo.
