+++ examples/minimal-react/client/package.json
@@ -4,7 +4,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
     "lint": "eslint --cache --ext \".js,.ts,.tsx\" src",
     "start": "vite preview"
   },
