+++ examples/minimal-react/client/src/main.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom/client';
 import { App } from './App';
 
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
