+++ examples/minimal-react/server/package.json
@@ -1,10 +1,11 @@
 {
   "name": "examples-minimal-react-server",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "tsc",
-    "dev": "tsx watch index.ts",
-    "lint": "eslint --cache --ext \".js,.ts,.tsx\" index.ts",
+    "build": "tsc && tsc-alias",
+    "dev": "tsx watch src/index.ts",
+    "lint": "eslint --cache --ext \".js,.ts,.tsx\" src",
     "start": "node dist/index.js"
   },
   "dependencies": {
@@ -16,6 +17,7 @@
     "@types/cors": "^2.8.13",
     "@types/node": "^20.10.0",
     "eslint": "^8.57.0",
+    "tsc-alias": "^1.8.10",
     "tsx": "^4.0.0",
     "typescript": "^5.6.2"
   }
