+++ examples/minimal-react/server/src/trpc.ts
@@ -10,5 +10,5 @@ const t = initTRPC.create();
  * Export reusable router and procedure helpers
  * that can be used throughout the router
  */
-export const router = t.router;
 export const publicProcedure = t.procedure;
+export const router = t.router;
