+++ examples/minimal-react/server/tsconfig.json
@@ -1,12 +1,14 @@
 {
+  "extends": "../tsconfig.json",
+  
+  "include": ["src"],
+  
   "compilerOptions": {
     "lib": ["es2023"],
     "module": "node16",
     "target": "es2022",
 
-    "strict": true,
     "esModuleInterop": true,
-    "skipLibCheck": true,
     "moduleResolution": "node16",
     "outDir": "dist"
   }
