--- examples/npm-workspaces/README.md
@@ -1,27 +1,0 @@
-# A npm workspaces tRPC example
-
-Requires node 18 (for global fetch).
-
-## Playing around
-
-```bash
-# install all workspaces dependencies & create workspaces links
-npm i
-```
-
-Within *client/* and *server/* respectively, run
-
-```bash
-npm run dev
-```
-
-Try editing the ts files to see the type checking in action :)
-
-## Building
-
-Within *client/* and *server/* respectively, run
-
-```bash
-npm run build
-npm run start
-```
