--- examples/npm-workspaces/client/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@examples-npm-workspaces/client",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "tsc",
-    "dev": "wait-port 3000 && tsx watch src/index.ts",
-    "start": "node dist/index.js"
-  },
-  "dependencies": {
-    "@trpc/client": "npm:@trpc/client@next"
-  },
-  "devDependencies": {
-    "@examples-npm-workspaces/server": "*",
-    "tsx": "^4.0.0",
-    "typescript": "^5.6.2",
-    "wait-port": "^1.0.1"
-  }
-}
