--- examples/npm-workspaces/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "@examples-npm-workspaces/root",
-  "private": true,
-  
-  "workspaces": [
-    "client",
-    "server"
-  ]
-}
