--- examples/npm-workspaces/server/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@examples-npm-workspaces/server",
-  "private": true,
-  "type": "module",
-  "types": "src/index.ts",
-  "scripts": {
-    "build": "tsc && tsc-alias",
-    "dev": "tsx watch src/index.ts",
-    "start": "node dist/index.js"
-  },
-  "dependencies": {
-    "@trpc/server": "npm:@trpc/server@next",
-    "zod": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.10.0",
-    "tsc-alias": "^1.8.10",
-    "tsx": "^4.0.0",
-    "typescript": "^5.6.2"
-  }
-}
